Create wrf_reproduce_grid.py
WRF save the output in a very strange format. I normally use the Lambert conformal projection when running the model, since it has the smallest distortions in mid-latitudes. However, this LCC grid is not stored in the output. Therefore, I decided to reproduce this grid myself. 

The code reads the projection parameters (which are stored as attributes in the wrfout files) and sets up a lambert conformal projection with these parameters. Then, it reads the grid definitions (number of grid cells and grid spacing in both x and y direction), and sets up a regular meshgrid centered around 0,0. Together, these components form a consistent set of coordinates.

The wrfoutput does not provide these coordinates. Instead, the netcdf files have grid cell indices as dimensions. To geo-reference your data, the output files provide auxiliary XLAT and XLONG variables, which are 2D, irregular arrays. Ugly. 

This script compares the reproduced grid with the XLAT and XLONG. I found an average mismatch of 0.5 m, and a maximum mismatch near the edge of the domain of 53 m. That is around 53 N.